Created utility class and moved createNotePreviewButton helper method to new util class as static method. Preview area and new voice note show up after TranscribeAudio endpoint returns a response
