+++ web/src/util/noteUtils.js
@@ -1,0 +1,31 @@
+export default class NoteUtils {
+    
+    /**
+     * static utility method to generate the note preview button for the note preview area.
+     * @param {*} note the note a button is being made for.
+     * @returns the note preview button with attached event listener.
+     */
+    static createNotePreviewButton(note) {
+        // Create button element, set button text to note title, and set values
+        let notePreviewButton = document.createElement("button");
+        notePreviewButton.className = "button";
+        notePreviewButton.id = "note-preview-button";
+        notePreviewButton.type = "button";
+        notePreviewButton.textContent = note.title;
+        notePreviewButton.noteTitle = note.title;
+        notePreviewButton.noteContent = note.content;
+        notePreviewButton.noteDateCreated = note.dateCreated;
+
+        // Create listener. When note preview is clicked, update primary note with note values
+        const primaryNoteTitle = document.querySelector(".primary-note-title");
+        const primaryNoteContent = document.querySelector(".primary-note-content");
+        const primaryNoteDateCreated = document.querySelector(".primary-note-date-created");
+        notePreviewButton.addEventListener("click", function (evt) {
+            primaryNoteTitle.textContent = evt.target.noteTitle;
+            primaryNoteContent.textContent = evt.target.noteContent;
+            primaryNoteDateCreated.textContent = evt.target.noteDateCreated;
+        });
+
+        return notePreviewButton;
+    }
+}
